Service logic moved to Application & Introducing Pod Struct
- Moved most of the `Service` logic to the `Application` since it’s not
sharing much with the upcoming `Pod` structure.
- Updating tests to reflect this change.
- The `ServiceUtil.createSpecializationForDefinition` does the
automatic class detection based on definition
