+++ src/js/structs/Pod.js
@@ -1,0 +1,5 @@
+import Service from './Service';
+
+module.exports = class Pod extends Service {
+
+};
