+++ src/js/structs/ServiceTree.js
@@ -1,4 +1,3 @@
-import Application from './Application';
 import Framework from './Framework';
 import HealthSorting from '../constants/HealthSorting';
 import HealthStatus from '../constants/HealthStatus';
@@ -46,13 +45,9 @@ module.exports = class ServiceTree extends Tree {
         );
       }
 
-      // Check the DCOS_PACKAGE_FRAMEWORK_NAME label to determine if the item
-      // should be converted to an Application or Framework instance.
-      if (item.labels && item.labels.DCOS_PACKAGE_FRAMEWORK_NAME) {
-        return new Framework(item);
-      } else {
-        return new Application(item);
-      }
+      // Create the appropriate service according to definition
+      return ServiceUtil.createSpecializationForDefinition(item);
+
     });
   }
