+++ src/js/utils/__tests__/ServiceUtil-test.js
@@ -2,10 +2,43 @@ jest.dontMock('../ServiceUtil');
 jest.dontMock('../../structs/Service');
 jest.dontMock('../../constants/VolumeConstants');
 
+const Application = require('../../structs/Application');
+const Framework = require('../../structs/Framework');
 const Service = require('../../structs/Service');
 const ServiceUtil = require('../ServiceUtil');
 
 describe('ServiceUtil', function () {
+  describe('#createSpecializationForDefinition', function () {
+    it('should correctly create Application instances', function () {
+      let instance = ServiceUtil.createSpecializationForDefinition({
+        id: '/test',
+        cmd: 'sleep 1000;',
+        cpus: null,
+        mem: null,
+        disk: null,
+        instances: null
+      });
+
+      expect(instance instanceof Application).toBeTruthy();
+    });
+
+    it('should correctly create Framework instances', function () {
+      let instance = ServiceUtil.createSpecializationForDefinition({
+        id: '/test',
+        cmd: 'sleep 1000;',
+        cpus: null,
+        mem: null,
+        disk: null,
+        instances: null,
+        labels: {
+          DCOS_PACKAGE_FRAMEWORK_NAME: 'Test Framework'
+        }
+      });
+
+      expect(instance instanceof Framework).toBeTruthy();
+    });
+  });
+
   describe('#createServiceFromFormModel', function () {
     it('should convert to the correct Service', function () {
       let model = {
@@ -15,7 +48,7 @@ describe('ServiceUtil', function () {
         }
       };
 
-      let expectedService = new Service({
+      let expectedService = new Application({
         id: '/test',
         cmd: 'sleep 1000;',
         cpus: null,
@@ -31,7 +64,7 @@ describe('ServiceUtil', function () {
     it('should return empty service if null is provided', function () {
       let model = null;
 
-      let expectedService = new Service({});
+      let expectedService = new Application({});
 
       expect(ServiceUtil.createServiceFromFormModel(model))
         .toEqual(expectedService);
@@ -40,7 +73,7 @@ describe('ServiceUtil', function () {
     it('should return empty service if empty object is provided', function () {
       let model = {};
 
-      let expectedService = new Service({});
+      let expectedService = new Application({});
 
       expect(ServiceUtil.createServiceFromFormModel(model))
         .toEqual(expectedService);
@@ -489,7 +522,7 @@ describe('ServiceUtil', function () {
           }
         };
 
-        let expectedService = new Service({
+        let expectedService = new Application({
           container: {
             type: 'MESOS',
             volumes: [{
@@ -529,7 +562,7 @@ describe('ServiceUtil', function () {
           }
         };
 
-        let expectedService = new Service({
+        let expectedService = new Application({
           container: {
             type: 'DOCKER',
             docker: {
@@ -575,7 +608,7 @@ describe('ServiceUtil', function () {
           }
         };
 
-        let expectedService = new Service({
+        let expectedService = new Application({
           container: {
             type: 'DOCKER',
             docker: {
@@ -605,7 +638,7 @@ describe('ServiceUtil', function () {
           }
         };
 
-        let expectedService = new Service({
+        let expectedService = new Application({
           container: {
             type: 'MESOS',
             volumes: [
@@ -689,7 +722,7 @@ describe('ServiceUtil', function () {
         ]
       };
 
-      let service = new Service({
+      let service = new Application({
         id: '/test',
         cmd: 'sleep 1000;'
       });
@@ -705,7 +738,7 @@ describe('ServiceUtil', function () {
 
   describe('#getAppDefinitionFromService', function () {
     it('should create the correct appDefinition', function () {
-      let service = new Service({
+      let service = new Application({
         id: '/test',
         cmd: 'sleep 1000;'
       });
@@ -719,7 +752,7 @@ describe('ServiceUtil', function () {
 
   describe('#convertServiceLabelsToArray', function () {
     it('should return an array of key-value tuples', function () {
-      let service = new Service({
+      let service = new Application({
         id: '/test',
         cmd: 'sleep 1000;',
         labels: {
@@ -737,7 +770,7 @@ describe('ServiceUtil', function () {
     });
 
     it('should return an empty array if no labels are found', function () {
-      let service = new Service({
+      let service = new Application({
         id: '/test',
         cmd: 'sleep 1000;'
       });
